refactor: reducers
