+++ app/src/store/slices/tabs-layout/reducers.ts
@@ -1,0 +1,51 @@
+import { CaseReducer, PayloadAction } from "@reduxjs/toolkit";
+import { TabsLayout } from "layouts/TabsLayout";
+
+export type TabsLayoutState = Record<
+  string,
+  {
+    tabs: TabsLayout.Tab[];
+    activeTab: TabsLayout.Tab | null;
+  }
+>;
+
+export const initialState: TabsLayoutState = {
+  apiClient: {
+    tabs: [],
+    activeTab: null,
+  },
+};
+
+export const resetState = () => initialState;
+
+export const addTab: CaseReducer<TabsLayoutState, PayloadAction<{ featureId: string; tab: TabsLayout.Tab }>> = (
+  state,
+  action
+) => {
+  const { featureId, tab } = action.payload;
+  state[featureId].tabs.push(tab);
+};
+
+export const updateTab: CaseReducer<
+  TabsLayoutState,
+  PayloadAction<{ featureId: string; tabId: TabsLayout.Tab["id"]; updatedTabData: Partial<TabsLayout.Tab> }>
+> = (state, action) => {
+  const { featureId, tabId, updatedTabData } = action.payload;
+  state[featureId].tabs = state[featureId].tabs.map((tab) => (tab.id === tabId ? { ...tab, ...updatedTabData } : tab));
+};
+
+export const setActiveTab: CaseReducer<TabsLayoutState, PayloadAction<{ featureId: string; tab: TabsLayout.Tab }>> = (
+  state,
+  action
+) => {
+  const { featureId, tab } = action.payload;
+  state[featureId].activeTab = { ...(state[featureId].activeTab ?? {}), ...tab };
+};
+
+export const setTabs: CaseReducer<TabsLayoutState, PayloadAction<{ featureId: string; tabs: TabsLayout.Tab[] }>> = (
+  state,
+  action
+) => {
+  const { featureId, tabs } = action.payload;
+  state[featureId].tabs = tabs;
+};
